requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
sqlalchemy>=2.0.9
alembic>=1.10.3
psycopg2-binary>=2.9.6
ccxt>=3.0.0
pandas>=2.0.0
numpy>=1.24.2
scikit-learn>=1.2.2
tensorflow>=2.12.0
stable-baselines3>=1.8.0
websockets>=11.0.2
aiohttp>=3.8.4
python-dotenv>=1.0.0
pytest>=7.3.1
httpx>=0.24.0
asyncio>=3.4.3
newsapi-python>=0.2.7
tweepy>=4.14.0
praw>=7.7.0
